+++ lib/src/graphics/point.dart
@@ -8,6 +8,7 @@ import '../os/parcelable.dart' show Parcelable;
 /// [Point] holds two integer coordinates.
 ///
 /// See: https://developer.android.com/reference/android/graphics/Point
+/// See: https://github.com/aosp-mirror/platform_frameworks_base/blob/master/graphics/java/android/graphics/Point.java
 class Point extends math.Point<int> with Parcelable {
   Point([final int x = 0, final int y = 0]) : super(x, y);
 
@@ -16,6 +17,7 @@ class Point extends math.Point<int> with Parcelable {
 
   @override
   void writeToParcel(final Parcel parcel, [final int flags = 0]) {
-    throw UnimplementedError(); // TODO
+    parcel.writeInt(x);
+    parcel.writeInt(y);
   }
 }