+++ lib/src/graphics/pointf.dart
@@ -8,6 +8,7 @@ import '../os/parcelable.dart' show Parcelable;
 /// [PointF] holds two float coordinates.
 ///
 /// See: https://developer.android.com/reference/android/graphics/PointF
+/// See: https://github.com/aosp-mirror/platform_frameworks_base/blob/master/graphics/java/android/graphics/PointF.java
 class PointF extends math.Point<double> with Parcelable {
   PointF([final double x = 0, final double y = 0]) : super(x, y);
 
@@ -16,6 +17,7 @@ class PointF extends math.Point<double> with Parcelable {
 
   @override
   void writeToParcel(final Parcel parcel, [final int flags = 0]) {
-    throw UnimplementedError(); // TODO
+    parcel.writeFloat(x);
+    parcel.writeFloat(y);
   }
 }